Refactor consensus candidate handling
